Send session info as POST request to backend.
Configured axios.post() call to make a successful POST request to the backend with a local port. Had to add @CrossOrigin(origins = "http://localhost:3000") because the URL of the front-end is different than the URL of the backend. Without this addition, browser will not process POST request because the server won't support a cross origin request (COR policy error).