Upload mentor attendance correctly.
Previously the mentor attendance was marked using the participants version of the URL. This was incorrect and it now uses the staff version. This resolves 2 issues Federica had in her feedback video. 1) Mentor attendance should appear under the staff section in Views, not participants. 2) Session attendance didn't appear under the mentor's page in Views > Session Attendances. Also, the Views database returns a 501 Not Implemented when using the staff version of the URL. But this is wrong, it is implemented and it does work, clearly, as it does what it says it does in the docs. So added logic to catch the 501 and ignore it when the POST request is sent.